Make project gradle-based for easy importing and test running.
