+++ .gitignore
@@ -1,0 +1,4 @@
+/build
+.gradle/
+.idea/
+*.iml
