+++ src/main/java/IntervalSetTree.java
@@ -1,40 +1,45 @@
-package datastructures;
-
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.NoSuchElementException;
-import java.util.Optional;
-import java.util.Set;
+import java.util.*;
 
 /**
  * A balanced binary-search tree keyed by Interval objects.
  * <p>
+ * This tree does not store exact duplicates, but will store Intervals that
+ * have identical coordinates but differ in some other aspect (e.g., a name
+ * field). Each Node of this tree stores its "identical" Intervals in a
+ * HashSet; hence the name "IntervalSetTree".
+ * <p>
+ * Two Intervals, i and j, will be stored as distinct Intervals in the same
+ * Node if and only if:
+ * <ul>
+ * <li>i.compareTo(j) == 0
+ * <li>i.equals(j) == false
+ * </ul><p>
  * The underlying data-structure is a red-black tree largely implemented from
  * CLRS (Introduction to Algorithms, 2nd edition) with the interval-tree
  * extensions mentioned in section 14.3
  * @param <I> - the type of Interval this tree contains
  */
-public class IntervalTree<T extends Interval> implements Iterable<T> {
+public class IntervalSetTree<T extends Interval> implements Iterable<T> {
 
     private Node root;  // The root Node.
     private Node nil;   // The sentinel Node to represent the absence of a node.
-    private int size;   // Size of the tree. Updated by insert() and Node#delete()
+    private int size;   // Size of the tree. Updated by insert() and delete()
 
     /**
-     * Constructs an empty IntervalTree.
+     * Constructs an empty IntervalSetTree.
      */
-    public IntervalTree() {
+    public IntervalSetTree() {
         nil = new Node();
         root = nil;
         size = 0;
     }
     
     /**
-     * Constructs an IntervalTree containing a single node corresponding to
-     * the given interval.
-     * @param t - the interval to add to the tree
+     * Constructs an IntervalSetTree with a single node containing the given
+     * Interval.
+     * @param t - the Interval to add to this IntervalSetTree
      */
-    public IntervalTree(T t) {
+    public IntervalSetTree(T t) {
         nil = new Node();
         root = new Node(t);
         root.blacken();
@@ -46,97 +51,102 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     ///////////////////////////////////
     
     /**
-     * Whether this IntervalTree is empty or not.
+     * Whether this IntervalSetTree is empty or not.
      */
     public boolean isEmpty() {
         return root.isNil();
     }
     
     /**
-     * The number of intervals stored in this IntervalTree.
+     * The number of intervals stored in this IntervalSetTree.
      */
     public int size() {
         return size;
     }
     
     /**
-     * The Node in this IntervalTree that contains the given Interval.
+     * The Node in this IntervalSetTree that has the same start and end
+     * coordinates of the given Interval.
      * <p>
-     * This method returns the nil Node if the Interval t cannot be found.
-     * @param t - the Interval to search for.
+     * It is not necessarily the case that the Node contains the Interval,
+     * just that the boundary coordinates are the same. This method returns
+     * the nil Node if no valid Node can be found.
+     * @param t - the Interval to search for
      */
     private Node search(T t) {
         return root.search(t);
     }
     
     /**
-     * Whether or not this IntervalTree contains the given Interval.
+     * Whether or not this IntervalSetTree contains the given Interval.
      * @param t - the Interval to search for
      */
     public boolean contains(T t) {
-        return !search(t).isNil();
+        return search(t).intervals.contains(t);
     }
     
     /**
-     * The minimum value in this IntervalTree
-     * @return an Optional containing, if it exists, the minimum value in this
-     * IntervalTree; otherwise (i.e., if this is empty), an empty Optional.
+     * The Intervals in the minimum Node of this IntervalSetTree.
+     * @return an Iterator, possibly empty, over all minimum Intervals
      */
-    public Optional<T> minimum() {
+    public Iterator<T> minimum() {
         Node n = root.minimumNode();
-        return n.isNil() ? Optional.empty() : Optional.of(n.interval());
+        return n.isNil() ? Collections.emptyIterator() : n.intervals.iterator();
     }
 
     /**
-     * The maximum value in this IntervalTree
-     * @return an Optional containing, if it exists, the maximum value in this
-     * IntervalTree; otherwise (i.e., if this is empty), an empty Optional.
+     * The Intervals in the maximum Node of this IntervalSetTree.
+     * @return an Iterator, possibly empty, over all maximum Intervals
      */
-    public Optional<T> maximum() {
+    public Iterator<T> maximum() {
         Node n = root.maximumNode();
-        return n.isNil() ? Optional.empty() : Optional.of(n.interval());
+        return n.isNil() ? Collections.emptyIterator() : n.intervals.iterator();
     }
     
     /**
-     * The next Interval in this IntervalTree
+     * The Intervals in the following Node of this IntervalSetTree.
+     * <p>
+     * The "following" Node is the next Node in this tree relative to the Node
+     * corresponding to the passed Interval
      * @param t - the Interval to search for
-     * @return an Optional containing, if it exists, the next Interval in this
-     * IntervalTree; otherwise (if t is the maximum Interval, or if this
-     * IntervalTree does not contain t), an empty Optional.
+     * @return an Iterator over the Intervals in the next Node, possibly empty
+     * if this Node is the maximum Node.
      */
-    public Optional<T> successor(T t) {
+    public Iterator<T> successors(T t) {
         Node n = search(t);
         if (n.isNil()) {
-            return Optional.empty();
+            return Collections.emptyIterator();
         }
 
         n = n.successor();
         if (n.isNil()) {
-            return Optional.empty();
+            return Collections.emptyIterator();
         }
 
-        return Optional.of(n.interval());
+        return n.intervals.iterator();
     }
 
     /**
-     * The previous Interval in this IntervalTree
+     * The Intervals in the preceding Node of this IntervalSetTree.
+     * <p>
+     * The "preceding" Node is the previous Node in this tree relative to the
+     * Node corresponding to the passed Interval
      * @param t - the Interval to search for
-     * @return an Optional containing, if it exists, the previous Interval in
-     * this IntervalTree; otherwise (if t is the minimum Interval, or if this
-     * IntervalTree does not contain t), an empty Optional.
+     * @return an Iterator over the Intervals in the next Node, possibly empty
+     * if this Node is the minimum Node.
      */
-    public Optional<T> predecessor(T t) {
+    public Iterator<T> predecessors(T t) {
         Node n = search(t);
         if (n.isNil()) {
-            return Optional.empty();
+            return Collections.emptyIterator();
         }
 
         n = n.predecessor();
         if (n.isNil()) {
-            return Optional.empty();
+            return Collections.emptyIterator();
         }
 
-        return Optional.of(n.interval());
+        return n.intervals.iterator();
     }
 
     /**
@@ -147,7 +157,7 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     }
     
     /**
-     * An Iterator over the Intervals in this IntervalTree that overlap the
+     * An Iterator over the Intervals in this IntervalSetTree that overlap the
      * given Interval
      * @param t - the overlapping Interval
      */
@@ -156,8 +166,8 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     }
     
     /**
-     * Whether or not any of the Intervals in this IntervalTree overlap the given
-     * Interval
+     * Whether or not any of the Intervals in this IntervalSetTree overlap the
+     * given Interval
      * @param t - the potentially overlapping Interval
      */
     public boolean overlaps(T t) {
@@ -165,24 +175,27 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     }
     
     /**
-     * The number of Intervals in this IntervalTree that overlap the given
+     * The number of Intervals in this IntervalSetTree that overlap the given
      * Interval
      * @param t - the overlapping Interval
      */
     public int numOverlappers(T t) {
-        return root.numOverlappingNodes(t);
+        return root.numOverlappingIntervals(t);
     }
     
     /**
-     * The least Interval in this IntervalTree that overlaps the given Interval
+     * The minimum Intervals in this IntervalSetTree that overlap the given
+     * Interval
+     * <p>
+     * There may be more than one minimum Interval if two Intervals have the same
+     * start and end coordinates
      * @param t - the overlapping Interval
-     * @return an Optional containing, if it exists, the least Interval in this
-     * IntervalTree that overlaps the given Interval; otherwise (i.e., if there
-     * is no overlap), an empty Optional
+     * @return a Iterator over the minimum Intervals that overlap the given
+     * Interval; an empty Iterator if no such Interval exists.
      */
-    public Optional<T> minimumOverlapper(T t) {
+    public Iterator<T> minimumOverlappers(T t) {
         Node n = root.minimumOverlappingNode(t);
-        return n.isNil() ? Optional.empty() : Optional.of(n.interval());
+        return n.isNil() ? Collections.emptyIterator() : n.intervals.iterator();
     }
     
     ///////////////////////////////
@@ -190,43 +203,57 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     ///////////////////////////////
 
     /**
-     * Inserts the given value into the IntervalTree.
+     * Inserts the given Interval into this IntervalSetTree.
      * <p>
-     * This method constructs a new Node containing the given value and places
-     * it into the tree. If the value already exists within the tree, the tree
-     * remains unchanged.
-     * @param t - the value to place into the tree
+     * If the Interval already exists within the tree, the tree remains
+     * unchanged. If the Interval doesn't exist, but has the same start and end
+     * coordinates as another contained Interval, it is placed within the tree
+     * as normal.
+     * @param t - the Interval to place into this tree
      * @return if the value did not already exist, i.e., true if the tree was
      * changed, false if it was not
      */
     public boolean insert(T t) {
         
-        Node z = new Node(t);
         Node y = nil;
         Node x = root;
 
-        while (!x.isNil()) {                         // Traverse the tree down to a leaf.
+        // Traverse the tree down to a leaf
+        while (!x.isNil()) {
             y = x;
-            x.maxEnd = Math.max(x.maxEnd, z.maxEnd); // Update maxEnd on the way down.
-            int cmp = z.compareTo(x);
+            
+            // Update maxEnd on the way down.
+            x.maxEnd = Math.max(x.maxEnd, t.end());
+
+            // If the Node for this Interval already exists, add the Interval
+            // to its Set and increment size if successful.
+            int cmp = t.compareTo(x);
             if (cmp == 0) {
-                return false;                        // Value already in tree. Do nothing.
+                if (x.intervals.add(t)) {
+                    size++;
+                    return true;
+                } else {
+                    return false;
+                }
             }
             x = cmp == -1 ? x.left : x.right;
         }
 
+        // Didn't find the correct Node on the way down, so make a new Node
+        // containing the Interval.
+        Node z = new Node(t);
         z.parent = y;
-       
-        if (y.isNil()) {
+
+        if (y.isNil()) {                // Three cases:
             root = z;
-            root.blacken();
-        } else {                      // Set the parent of n.
+            root.blacken();             // 1) New node is root
+        } else {
             int cmp = z.compareTo(y);
             if (cmp == -1) {
-                y.left = z;
+                y.left = z;             // 2) New node is left-child leaf
             } else {
                 assert(cmp == 1);
-                y.right = z;
+                y.right = z;            // 3) New node is right-child leaf
             }
             
             z.left = nil;
@@ -234,7 +261,6 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
             z.redden();
             z.insertFixup();
         }
-
         size++;
         return true;
     }
@@ -244,33 +270,57 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     //////////////////////////////
     
     /**
-     * Deletes the given value from this IntervalTree.
+     * Deletes the given Interval from this IntervalSetTree.
      * <p>
-     * If the value does not exist, this IntervalTree remains unchanged. 
+     * If the Interval does not exist, this IntervalTree remains unchanged. 
      * @param t - the Interval to delete from the tree
-     * @return whether or not an Interval was removed from this IntervalTree
+     * @return whether or not an Interval was removed from this tree
      */
-    public boolean delete(T t) {    // Node#delete does nothing and returns
-        return search(t).delete();  // false if t.isNil()
+    public boolean delete(T t) {
+        Node n = search(t);
+        
+        boolean rtrn = n.intervals.remove(t);
+
+        if (rtrn) {
+            size--;
+        }
+
+        if (n.intervals.isEmpty()) {
+            n.delete(); // Node#delete does nothing if n.isNil()
+        }
+
+        return rtrn;
     }
     
     /**
-     * Deletes the smallest Interval from this IntervalTree.
+     * Deletes all Intervals from this IntervalSetTree that have the same start
+     * and end coordinates as the given Interval
+     * @param t - the Interval to delete
+     * @return whether or not any Intervals were removed from this tree
+     */
+    public boolean deleteSameBounds(T t) {
+        return search(t).delete();
+    }
+    
+    /**
+     * Deletes the smallest Intervals from this IntervalSetTree.
      * <p>
      * If there is no smallest Interval (that is, if the tree is empty), this
-     * IntervalTree remains unchanged.
-     * @return whether or not an Interval was removed from this IntervalTree
+     * IntervalSetTree remains unchanged. If multiple Intervals share the same
+     * start and end value, all are removed.
+     * @return whether or not any Intervals were removed from this tree
      */
     public boolean deleteMin() {            // Node#delete does nothing and
         return root.minimumNode().delete(); // returns false if t.isNil()
     }
     
     /**
-     * Deletes the greatest Interval from this IntervalTree.
+     * Deletes the greatest Intervals from this IntervalSetTree.
      * <p>
      * If there is no greatest Interval (that is, if the tree is empty), this
-     * IntervalTree remains unchanged.
-     * @return whether or not an Interval was removed from this IntervalTree
+     * IntervalSetTree remains unchanged. If multiple Intervals share the same
+     * start and end value, all are removed.
+     * @return whether or not any Intervals were removed from this tree
      */
     public boolean deleteMax() {            // Node#delete does nothing and
         return root.maximumNode().delete(); // returns false if t.isNil()
@@ -278,12 +328,12 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     
     /**
      * Deletes all Intervals that overlap the given Interval from this
-     * IntervalTree.
+     * IntervalSetTree.
      * <p>
-     * If there are no overlapping Intervals, this IntervalTree remains
+     * If there are no overlapping Intervals, this IntervalSetTree remains
      * unchanged.
      * @param t - the overlapping Interval
-     * @return whether or not an Interval was removed from this IntervalTree
+     * @return whether or not any Interval were removed from this tree
      */
     public boolean deleteOverlappers(T t) {
         // TODO 
@@ -300,21 +350,23 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
         Iterator<Node> iter = new OverlappingNodeIterator(root, t);
         iter.forEachRemaining(s::add);
         return s.stream()
-                .map(n -> delete(n.interval))
+                .map(n -> deleteSameBounds(n.intervals.iterator().next()))
                 .reduce(false, (a, b) -> a || b);
     }
 
     /**
      * A representation of a node in an interval tree.
      */
-    private class Node implements Interval {
+    private class Node implements Interval, Iterable<T> {
         
         /* Most of the "guts" of the interval tree are actually methods called
          * by nodes. For example, IntervalTree#delete(val) searches up the Node
          * containing val; then that Node deletes itself with Node#delete().
          */
 
-        private T interval;
+        private Set<T> intervals;
+        private int start;
+        private int end;
         private Node parent;
         private Node left;
         private Node right;
@@ -324,11 +376,12 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
         /**
          * Constructs a Node with no data.
          * <p>
-         * This Node has a null interval field, is black, and has all pointers
+         * This Node contains no Intervals, is black, and has all pointers
          * pointing at itself. This is intended to be used as the sentinel
          * node in the tree ("nil" in CLRS).
          */
         private Node() {
+            intervals = Collections.emptySet();
             parent = this;
             left = this;
             right = this;
@@ -340,35 +393,35 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
          * @param data - the Interval to be contained within this Node
          */
         public Node(T interval) {
-            this.interval = interval;
+            intervals = new HashSet<T>();
+            intervals.add(interval);
             parent = nil;
             left = nil;
             right = nil;
-            maxEnd = interval.end();
+            start = interval.start();
+            end = interval.end();
+            maxEnd = end;
             redden();
         }
         
-        /**
-         * The Interval in this Node
-         */
-        public T interval() {
-            return interval;
+        public Iterator<T> iterator() {
+            return intervals.iterator();
         }
         
         /**
-         * The start of the Interval in this Node
+         * The start of the Intervals in this Node
          */
         @Override
         public int start() {
-            return interval.start();
+            return start;
         }
 
         /**
-         * The end of the Interval in this Node
+         * The end of the Intervals in this Node
          */
         @Override
         public int end() {
-            return interval.end();
+            return end;
         }
         
         ///////////////////////////////////
@@ -376,9 +429,14 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
         ///////////////////////////////////
         
         /**
-         * Searches the subtree rooted at this Node for the given Interval.
+         * Searches the subtree rooted at this Node for the Node with the
+         * coordinates represented by this Interval.
+         * <p>
+         * The Interval does not need to be contained within the Node, just
+         * have the same coordinates, to be returned.
          * @param t - the Interval to search for
-         * @return the Node with the given Interval, if it exists; otherwise,
+         * @return the Node with the corresponding coordinates, if it exists;
+         * otherwise,
          * the sentinel Node 
          */
         private Node search(T t) {
@@ -392,8 +450,8 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
         }
 
         /**
-         * Searches the subtree rooted at this Node for its minimum Interval.
-         * @return the Node with the minimum Interval, if it exists; otherwise,
+         * Searches the subtree rooted at this Node for its minimum Intervals.
+         * @return the Node with the minimum Intervals, if it exists; otherwise,
          * the sentinel Node
          */
         private Node minimumNode() {
@@ -407,8 +465,8 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
         }
 
         /**
-         * Searches the subtree rooted at this Node for its maximum Interval.
-         * @return the Node with the maximum Interval, if it exists; otherwise,
+         * Searches the subtree rooted at this Node for its maximum Intervals.
+         * @return the Node with the maximum Intervals, if it exists; otherwise,
          * the sentinel Node
          */
         private Node maximumNode() {
@@ -474,7 +532,7 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
          * The only guarantee of this method is that the returned Node overlaps
          * the Interval t. This method is meant to be a quick helper method to
          * determine if any overlap exists between an Interval and any of an
-         * IntervalTree's Intervals. The returned Node will be the first
+         * IntervalSetTree's Intervals. The returned Node will be the first
          * overlapping one found.
          * @param t - the given Interval
          * @return an overlapping Node from this Node's subtree, if one exists;
@@ -482,7 +540,7 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
          */
         private Node anyOverlappingNode(T t) {
             Node x = this;
-            while (!x.isNil() && !t.overlaps(x.interval)) {
+            while (!x.isNil() && !t.overlaps(x)) {
                 x = !x.left.isNil() && x.left.maxEnd > t.start() ? x.left : x.right;
             }
             return x;
@@ -584,23 +642,22 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
         }
         
         /**
-         * The number of Nodes in this Node's subtree that overlap the given
+         * The number of Intervals in this Node's subtree that overlap the given
          * Interval t.
          * <p>
-         * This number includes this Node if this Node overlaps t. This method
-         * iterates over all overlapping Nodes, so if you ultimately need to
-         * inspect the Nodes, it will be more efficient to simply create the
-         * Iterator yourself.
+         * This number includes this Node's Intervals if they overlap t. This
+         * method iterates over all overlapping Nodes, so if you ultimately
+         * need to inspect the Intervals, it will be more efficient to simply
+         * create the Iterator yourself.
          * @param t - the overlapping Interval
          * @return the number of overlapping Nodes
          */
-        private int numOverlappingNodes(T t) {
+        private int numOverlappingIntervals(T t) {
             int count = 0;
             Iterator<Node> iter = new OverlappingNodeIterator(this, t);
             
             while (iter.hasNext()) {
-                iter.next();
-                count++;
+                count += iter.next().intervals.size();
             }
             return count;
         }
@@ -628,6 +685,7 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
                 return false;
             }
             
+            size -= this.intervals.size();
             Node y = this;
 
             if (hasTwoChildren()) { // If the node to remove has two children,
@@ -656,7 +714,6 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
                 x.deleteFixup();
             }
             
-            size--;
             return true;
         }
         
@@ -748,7 +805,7 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
          * correct maxEnd values.
          */
         private void resetMaxEnd() {
-            int val = interval.end();
+            int val = end;
             if (!left.isNil()) {
                 val = Math.max(val, left.maxEnd);
             }
@@ -834,9 +891,18 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
          * @param o - the other Node containing the data to be copied
          */
         private void copyData(Node o) {
-            interval = o.interval;
+            intervals = o.intervals;
+            start = o.start;
+            end = o.end;
         }
         
+        /**
+         * Returns a String representation of this Node.
+         * <p>
+         * This representation will display the start and end coordinates, the
+         * color, and the max-end value of this Node. Useful for quick
+         * debugging outside of the debugger.
+         */
         @Override
         public String toString() {
             if (isNil()) {
@@ -1056,8 +1122,9 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     ///////////////////////
     
     /**
-     * An Iterator which walks along this IntervalTree's Nodes in ascending order.
+     * An Iterator which walks along this IntervalSetTree's Nodes in ascending order.
      */
+    @SuppressWarnings("unused")
     private class TreeNodeIterator implements Iterator<Node> {
 
         private Node next;
@@ -1083,32 +1150,41 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     }
 
     /**
-     * An Iterator which walks along this IntervalTree's Intervals in ascending
-     * order.
-     * <p>
-     * This class just wraps a TreeNodeIterator and extracts each Node's Interval.
+     * An Iterator which walks along this IntervalSetTree's Intervals in
+     * ascending order.
      */
     private class TreeIterator implements Iterator<T> {
         
-        private TreeNodeIterator nodeIter;
+        private Iterator<T> iter;
+        private Node currentNode;
+        private Node nextNode;
         
         private TreeIterator(Node root) {
-            nodeIter = new TreeNodeIterator(root);
+            currentNode = root.minimumNode();
+            nextNode = currentNode.successor();
+            iter = currentNode.iterator();
         }
 
         @Override
         public boolean hasNext() {
-            return nodeIter.hasNext();
+            return iter.hasNext() || !nextNode.isNil();
         }
 
         @Override
         public T next() {
-            return nodeIter.next().interval;
+            if (iter.hasNext()) {
+                return iter.next();
+            } else {
+                currentNode = nextNode;
+                nextNode = currentNode.successor();
+                iter = currentNode.iterator();
+                return iter.next();
+            }
         }
     }
  
     /**
-     * An Iterator which walks along this IntervalTree's Nodes that overlap
+     * An Iterator which walks along this IntervalSetTree's Nodes that overlap
      * a given Interval in ascending order.
      */
     private class OverlappingNodeIterator implements Iterator<Node> {
@@ -1138,28 +1214,38 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     }
 
     /**
-     * An Iterator which walks along this IntervalTree's Intervals that overlap
+     * An Iterator which walks along this IntervalSetTree's Intervals that overlap
      * a given Interval in ascending order.
-     * <p>
-     * This class just wraps an OverlappingNodeIterator and extracts each Node's
-     * Interval.
      */
     private class OverlapperIterator implements Iterator<T> {
         
-        private OverlappingNodeIterator nodeIter;
+        private Iterator<T> iter;
+        private Node currentNode;
+        private Node nextNode;
+        private T t;
         
         private OverlapperIterator(Node root, T t) {
-            nodeIter = new OverlappingNodeIterator(root, t);
+            this.t = t;
+            currentNode = root.minimumOverlappingNode(t);
+            nextNode = currentNode.nextOverlappingNode(t);
+            iter = currentNode.iterator();
         }
 
         @Override
         public boolean hasNext() {
-            return nodeIter.hasNext();
+            return iter.hasNext() || !nextNode.isNil();
         }
 
         @Override
         public T next() {
-            return nodeIter.next().interval;
+            if (iter.hasNext()) {
+                return iter.next();
+            } else {
+                currentNode = nextNode;
+                nextNode = currentNode.nextOverlappingNode(t);
+                iter = currentNode.iterator();
+                return iter.next();
+            }
         }
     }
 
@@ -1168,7 +1254,7 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     ///////////////////////////////
     
     /**
-     * Whether or not this IntervalTree is a valid binary-search tree.
+     * Whether or not this IntervalSetTree is a valid binary-search tree.
      * <p>
      * This method will return false if any Node is less than its left child
      * or greater than its right child.
@@ -1182,7 +1268,7 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     }
 
     /**
-     * Whether or not this IntervalTree is balanced.
+     * Whether or not this IntervalSetTree is balanced.
      * <p>
      * This method will return false if all of the branches (from root to leaf)
      * do not contain the same number of black nodes. (Specifically, the
@@ -1206,7 +1292,7 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     }
     
     /**
-     * Whether or not this IntervalTree has a valid red coloring.
+     * Whether or not this IntervalSetTree has a valid red coloring.
      * <p>
      * This method will return false if all of the branches (from root to leaf)
      * do not contain the same number of black nodes. (Specifically, the
@@ -1222,7 +1308,7 @@ public class IntervalTree<T extends Interval> implements Iterable<T> {
     }
     
     /**
-     * Whether or not this IntervalTree has consistent maxEnd values.
+     * Whether or not this IntervalSetTree has consistent maxEnd values.
      * <p>
      * This method will only return true if each Node has a maxEnd value equal
      * to the highest interval end value of all the intervals in its subtree.
